fix(env.example): fix db_name (#3)
* fix(env.example): fix db_name

* feat(db): implement initial schema for recipes, tags, and groups

This commit introduces the foundational database schema required for the core functionality of the recipe bot. It adds tables for managing recipes, tags, and the relationships between users and groups.

Key changes include:
- Creating the `recipes` table to store recipe details, including a new `recipe_category` ENUM type.
- Adding the `recipe_tag` table to allow for recipe categorization by users or groups.
- Implementing a `users_groups` junction table to establish a many-to-many relationship between users and groups.
- Refactoring the `groups` table to use the Telegram chat ID (`tg_chat_id`) as the primary key, simplifying its structure.
- Refining the `users` table by replacing `TEXT` with more specific `VARCHAR` types for better data integrity.
- Updating the table creation logic to execute the new SQL scripts upon initialization.

* chore(docker): Ensure API waits for healthy postgres service

The previous `depends_on` configuration only waited for the postgres container to start, not for the database within it to be ready to accept connections. This could lead to a race condition where the API service would fail to start because it couldn't connect to the database.

This change updates the dependency to use `condition: service_healthy`. This ensures the `recipe_api` service will only start after the `postgres_recipe` service has passed its health check and is fully operational, making the application startup more robust and reliable.

* feat(repository): Implement repository for recipe persistence

This commit introduces the persistence layer for recipes, following the hexagonal architecture pattern. A new `RecipeRepositoryABC` port has been created, along with its corresponding SQL adapter `RecipeAsyncpgRepo`. This allows for saving new recipes to the database.

Additionally, this commit includes several refactoring improvements to the existing `UserAsyncpgRepo`:
- Renamed `get_by_tg_id` to a more generic `get` for clarity.
- Replaced manual model instantiation `User(**dict(row))` with the more robust `User.model_validate(row)` for better type validation and coercion.
- Removed unnecessary logging and error checks.

* cfg: add test/prod bots to config

* WIP

* feat: handle basic CRUD, improve user flow

* feat: edit recipe handler

* feat: delete recipe handler

* ref: CRUD logic ref

* feat: list pagination

* feat: tags

* feat: fix inline keyvoards bugs

* fix: small UI bugs

* feat: tt description parsing

* feat: basic from_tiktok done

* feat: add source link saving

---------

Co-authored-by: Alex Zakrevsky <aliaksandr.zakreuski@hudsonandhayes.co.uk>